pyproject.toml
[project]
name = "prices"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gunicorn>=23.0.0",
    "jinja2>=3.1.5",
    "pymongo>=4.11.1",
    "python-telegram-bot>=21.11.1",
    "requests>=2.32.3",
    "simple-term-menu>=1.6.6",
    "tabulate>=0.9.0",
    "thefuzz>=0.22.1",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "flask>=3.1.0",
]

#[project.scripts]
#scrape = "prices:scrape:main"
#report = "prices:report:main"
#
#[build-system]
#requires = ["hatchling"]
#build-backend = "hatchling.build"
